refactoring postcode controller
